+++ crates/nostr-sqlite/src/lib.rs
@@ -20,8 +20,8 @@ use async_trait::async_trait;
 use nostr::nips::nip01::Coordinate;
 use nostr::{Event, EventId, Filter, Timestamp, Url};
 use nostr_database::{
-    Backend, DatabaseIndexes, EventIndexResult, FlatBufferBuilder, FlatBufferDecode,
-    FlatBufferEncode, NostrDatabase, Order, TempEvent,
+    Backend, DatabaseEventResult, DatabaseHelper, FlatBufferBuilder, FlatBufferDecode,
+    FlatBufferEncode, NostrDatabase, Order,
 };
 use rusqlite::config::DbConfig;
 use rusqlite::Connection;
@@ -39,13 +39,12 @@ use self::pool::Pool;
 #[derive(Debug, Clone)]
 pub struct SQLiteDatabase {
     pool: Pool,
-    indexes: DatabaseIndexes,
+    helper: DatabaseHelper,
     fbb: Arc<RwLock<FlatBufferBuilder<'static>>>,
 }
 
 impl SQLiteDatabase {
-    /// Open SQLite store
-    pub async fn open<P>(path: P) -> Result<Self, Error>
+    async fn new<P>(path: P, helper: DatabaseHelper) -> Result<Self, Error>
     where
         P: AsRef<Path>,
     {
@@ -57,35 +56,55 @@ impl SQLiteDatabase {
 
         let this = Self {
             pool,
-            indexes: DatabaseIndexes::new(),
+            helper,
             fbb: Arc::new(RwLock::new(FlatBufferBuilder::with_capacity(70_000))),
         };
 
-        // Build indexes
-        this.build_indexes().await?;
+        this.bulk_load().await?;
 
         Ok(this)
     }
 
+    /// Open database with **unlimited** capacity
+    #[inline]
+    pub async fn open<P>(path: P) -> Result<Self, Error>
+    where
+        P: AsRef<Path>,
+    {
+        Self::new(path, DatabaseHelper::unbounded()).await
+    }
+
+    /// Open database with **limited** capacity
+    #[inline]
+    pub async fn open_bounded<P>(path: P, max_capacity: usize) -> Result<Self, Error>
+    where
+        P: AsRef<Path>,
+    {
+        Self::new(path, DatabaseHelper::bounded(max_capacity)).await
+    }
+
     #[tracing::instrument(skip_all)]
-    async fn build_indexes(&self) -> Result<(), Error> {
+    async fn bulk_load(&self) -> Result<(), Error> {
         let events = self
             .pool
             .interact(move |conn| {
-                let mut stmt = conn.prepare_cached("SELECT event FROM events;")?;
+                // Query
+                let mut stmt = conn.prepare("SELECT event FROM events;")?;
                 let mut rows = stmt.query([])?;
+
+                // Decode
                 let mut events = BTreeSet::new();
                 while let Ok(Some(row)) = rows.next() {
                     let buf: &[u8] = row.get_ref(0)?.as_bytes()?;
-                    let raw = TempEvent::decode(buf)?;
-                    events.insert(raw);
+                    let event = Event::decode(buf)?;
+                    events.insert(event);
                 }
-                Ok::<BTreeSet<TempEvent>, Error>(events)
+                Ok::<BTreeSet<Event>, Error>(events)
             })
             .await??;
 
         // Build indexes
-        let to_discard: HashSet<EventId> = self.indexes.bulk_index(events).await;
+        let to_discard: HashSet<EventId> = self.helper.bulk_load(events).await;
 
         // Discard events
         if !to_discard.is_empty() {
@@ -114,10 +133,10 @@ impl NostrDatabase for SQLiteDatabase {
     #[tracing::instrument(skip_all, level = "trace")]
     async fn save_event(&self, event: &Event) -> Result<bool, Self::Err> {
         // Index event
-        let EventIndexResult {
+        let DatabaseEventResult {
             to_store,
             to_discard,
-        } = self.indexes.index_event(event).await;
+        } = self.helper.index_event(event).await;
 
         if !to_discard.is_empty() {
             self.pool
@@ -161,7 +180,7 @@ impl NostrDatabase for SQLiteDatabase {
         let mut fbb = self.fbb.write().await;
 
         // Events to store
-        let events = self.indexes.bulk_import(events).await;
+        let events = self.helper.bulk_import(events).await;
 
         // Encode
         let events: Vec<(EventId, Vec<u8>)> = events
@@ -193,7 +212,7 @@ impl NostrDatabase for SQLiteDatabase {
     }
 
     async fn has_event_already_been_saved(&self, event_id: &EventId) -> Result<bool, Self::Err> {
-        if self.indexes.has_event_id_been_deleted(event_id).await {
+        if self.helper.has_event_id_been_deleted(event_id).await {
             Ok(true)
         } else {
             let event_id: String = event_id.to_hex();
@@ -231,7 +250,7 @@ impl NostrDatabase for SQLiteDatabase {
     }
 
     async fn has_event_id_been_deleted(&self, event_id: &EventId) -> Result<bool, Self::Err> {
-        Ok(self.indexes.has_event_id_been_deleted(event_id).await)
+        Ok(self.helper.has_event_id_been_deleted(event_id).await)
     }
 
     async fn has_coordinate_been_deleted(
@@ -240,7 +259,7 @@ impl NostrDatabase for SQLiteDatabase {
         timestamp: Timestamp,
     ) -> Result<bool, Self::Err> {
         Ok(self
-            .indexes
+            .helper
             .has_coordinate_been_deleted(coordinate, timestamp)
             .await)
     }
@@ -296,36 +315,13 @@ impl NostrDatabase for SQLiteDatabase {
     #[inline]
     #[tracing::instrument(skip_all, level = "trace")]
     async fn count(&self, filters: Vec<Filter>) -> Result<usize, Self::Err> {
-        Ok(self.indexes.count(filters).await)
-    }
-
-    #[tracing::instrument(skip_all)]
-    async fn query(&self, filters: Vec<Filter>, order: Order) -> Result<Vec<Event>, Self::Err> {
-        let ids: Vec<EventId> = self.indexes.query(filters, order).await;
-        self.pool
-            .interact(move |conn| {
-                let mut events: Vec<Event> = Vec::with_capacity(ids.len());
-                let mut stmt =
-                    conn.prepare_cached("SELECT event FROM events WHERE event_id = ?;")?;
-                for id in ids.into_iter() {
-                    let mut rows = stmt.query([id.to_hex()])?;
-                    while let Ok(Some(row)) = rows.next() {
-                        let buf: &[u8] = row.get_ref(0)?.as_bytes()?;
-                        events.push(Event::decode(buf)?);
-                    }
-                }
-                Ok(events)
-            })
-            .await?
+        Ok(self.helper.count(filters).await)
     }
 
     #[inline]
-    async fn event_ids_by_filters(
-        &self,
-        filters: Vec<Filter>,
-        order: Order,
-    ) -> Result<Vec<EventId>, Self::Err> {
-        Ok(self.indexes.query(filters, order).await)
+    #[tracing::instrument(skip_all)]
+    async fn query(&self, filters: Vec<Filter>, order: Order) -> Result<Vec<Event>, Self::Err> {
+        Ok(self.helper.query(filters, order).await)
     }
 
     #[inline]
@@ -333,11 +329,11 @@ impl NostrDatabase for SQLiteDatabase {
         &self,
         filter: Filter,
     ) -> Result<Vec<(EventId, Timestamp)>, Self::Err> {
-        Ok(self.indexes.negentropy_items(filter).await)
+        Ok(self.helper.negentropy_items(filter).await)
     }
 
     async fn delete(&self, filter: Filter) -> Result<(), Self::Err> {
-        match self.indexes.delete(filter).await {
+        match self.helper.delete(filter).await {
             Some(ids) => {
                 self.pool
                     .interact(move |conn| {
@@ -377,7 +373,7 @@ impl NostrDatabase for SQLiteDatabase {
 
         migration::run(&self.pool).await?;
 
-        self.indexes.clear().await;
+        self.helper.clear().await;
 
         Ok(())
     }